working killf() and spawn()
Had to rename kill() to killf(), because of a conflict with another name (maybe in stdlib?).
Tweaked the makefile to work properly. Added defintions for the field and columns and rows.
New file, body.c, to contain all of the functions from dict.h.
Updated test.c and main.c appropriately.