Fixed generator bug: cells are not killed until entire field is scanned
Before, cells were being killed or spawned as they were looked at by rule(). I
added a list of function pointer that saves up the work that needs to be done.
That way, after each cell has been looked at, *then* the actions can be
performed. This avoid needing to make a copy of the field, which I think is
good.

Also cleaned up style and formatting, etc. Added debugging to the Makefile.